# src/main/java/com/eaglebank/user/api/account/dto/UpdateBankAccountRequest.java
package com.eaglebank.user.api.account.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;


public record UpdateBankAccountRequest(@NotBlank String name, @NotNull AccountType accountType)
{
}