# src/main/java/com/eaglebank/user/api/user/dto/LoginRequest.java
package com.eaglebank.user.api.user.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;


public record LoginRequest(@Email @NotBlank String email, @NotBlank String password)
{
}